requirements.txt
hypothesis
types-pillow
numpy-stubs